Struct stm32f1xx_hal::dma::CircBuffer

pub struct CircBuffer<BUFFER, PAYLOAD> where
    BUFFER: 'static,  { /* fields omitted */ }

Implementations

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C1>> where
RxDma<PAYLOAD, C1>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C1>)

Stops the transfer and returns the underlying buffer and RxDma

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C2>> where
RxDma<PAYLOAD, C2>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C2>)

Stops the transfer and returns the underlying buffer and RxDma

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C3>> where
RxDma<PAYLOAD, C3>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C3>)

Stops the transfer and returns the underlying buffer and RxDma

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C4>> where
RxDma<PAYLOAD, C4>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C4>)

Stops the transfer and returns the underlying buffer and RxDma

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C5>> where
RxDma<PAYLOAD, C5>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C5>)

Stops the transfer and returns the underlying buffer and RxDma

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C6>> where
RxDma<PAYLOAD, C6>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C6>)

Stops the transfer and returns the underlying buffer and RxDma

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C7>> where
RxDma<PAYLOAD, C7>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C7>)

Stops the transfer and returns the underlying buffer and RxDma

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C1>> where
RxDma<PAYLOAD, C1>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C1>)

Stops the transfer and returns the underlying buffer and RxDma

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C2>> where
RxDma<PAYLOAD, C2>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C2>)

Stops the transfer and returns the underlying buffer and RxDma

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C3>> where
RxDma<PAYLOAD, C3>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C3>)

Stops the transfer and returns the underlying buffer and RxDma

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C4>> where
RxDma<PAYLOAD, C4>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C4>)

Stops the transfer and returns the underlying buffer and RxDma

impl<B, PAYLOAD> CircBuffer<B, RxDma<PAYLOAD, C5>> where
RxDma<PAYLOAD, C5>: TransferPayload,

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

pub fn stop(self) -> (&'static mut [B; 2], RxDma<PAYLOAD, C5>)

Stops the transfer and returns the underlying buffer and RxDma