Struct stm32f1xx_hal::gpio::gpioa::PA6

pub struct PA6<MODE> { /* fields omitted */ }

Pin

Implementations

impl<MODE> PA6<MODE> where
MODE: Active,

pub fn into_alternate_push_pull(self, cr: &mut CRL) -> PA6<Alternate<PushPull>>

Configures the pin to operate as an alternate function push-pull output pin.

pub fn into_alternate_open_drain(
self,
cr: &mut CRL
) -> PA6<Alternate<OpenDrain>>

Configures the pin to operate as an alternate function open-drain output pin.

pub fn into_floating_input(self, cr: &mut CRL) -> PA6<Input<Floating>>

Configures the pin to operate as a floating input pin

pub fn into_pull_down_input(self, cr: &mut CRL) -> PA6<Input<PullDown>>

Configures the pin to operate as a pulled down input pin

pub fn into_pull_up_input(self, cr: &mut CRL) -> PA6<Input<PullUp>>

Configures the pin to operate as a pulled up input pin

pub fn into_open_drain_output(self, cr: &mut CRL) -> PA6<Output<OpenDrain>>

Configures the pin to operate as an open-drain output pin. Initial state will be low.

pub fn into_open_drain_output_with_state(
self,
cr: &mut CRL,
initial_state: State
) -> PA6<Output<OpenDrain>>

Configures the pin to operate as an open-drain output pin. initial_state specifies whether the pin should be initially high or low.

pub fn into_push_pull_output(self, cr: &mut CRL) -> PA6<Output<PushPull>>

Configures the pin to operate as an push-pull output pin. Initial state will be low.

pub fn into_push_pull_output_with_state(
self,
cr: &mut CRL,
initial_state: State
) -> PA6<Output<PushPull>>

Configures the pin to operate as an push-pull output pin. initial_state specifies whether the pin should be initially high or low.

pub fn into_analog(self, cr: &mut CRL) -> PA6<Analog>

Configures the pin to operate as an analog input pin

pub fn into_dynamic(self, cr: &mut CRL) -> PA6<Dynamic>

Configures the pin as a pin that can change between input and output without changing the type. It starts out as a floating input

impl<MODE> PA6<MODE> where
MODE: Active,
PA6<MODE>: PinMode<CRL>,

pub fn as_push_pull_output(
&mut self,
cr: &mut CRL,
f: impl FnMut(&mut PA6<Output<PushPull>>)
)

Temporarily change the mode of the pin.

The value of the pin after conversion is undefined. If you want to control it, use $stateful_fn_name

pub fn as_push_pull_output_with_state(
&mut self,
cr: &mut CRL,
state: State,
f: impl FnMut(&mut PA6<Output<PushPull>>)
)

Temporarily change the mode of the pin.

Note that the new state is set slightly before conversion happens. This can cause a short output glitch if switching between output modes

pub fn as_open_drain_output(
&mut self,
cr: &mut CRL,
f: impl FnMut(&mut PA6<Output<OpenDrain>>)
)

Temporarily change the mode of the pin.

The value of the pin after conversion is undefined. If you want to control it, use $stateful_fn_name

pub fn as_open_drain_output_with_state(
&mut self,
cr: &mut CRL,
state: State,
f: impl FnMut(&mut PA6<Output<OpenDrain>>)
)

Temporarily change the mode of the pin.

Note that the new state is set slightly before conversion happens. This can cause a short output glitch if switching between output modes

pub fn as_floating_input(
&mut self,
cr: &mut CRL,
f: impl FnMut(&mut PA6<Input<Floating>>)
)

Temporarily change the mode of the pin.

pub fn as_pull_up_input(
&mut self,
cr: &mut CRL,
f: impl FnMut(&mut PA6<Input<PullUp>>)
)

Temporarily change the mode of the pin.

pub fn as_pull_down_input(
&mut self,
cr: &mut CRL,
f: impl FnMut(&mut PA6<Input<PullDown>>)
)

Temporarily change the mode of the pin.

impl<MODE> PA6<MODE> where
MODE: Active,

pub fn downgrade(self) -> Pxx<MODE>

Erases the pin number and port from the type

This is useful when you want to collect the pins into an array where you need all the elements to have the same type

impl PA6<Dynamic>

pub fn make_pull_up_input(&mut self, cr: &mut CRL)

pub fn make_pull_down_input(&mut self, cr: &mut CRL)

pub fn make_floating_input(&mut self, cr: &mut CRL)

pub fn make_push_pull_output(&mut self, cr: &mut CRL)

pub fn make_open_drain_output(&mut self, cr: &mut CRL)