Trait stm32f1xx_hal::gpio::OutputSpeed

pub trait OutputSpeed<CR> {
    pub fn set_speed(&mut self, cr: &mut CR, speed: IOPinSpeed);
}

Allow setting of the slew rate of an IO pin

Initially all pins are set to the maximum slew rate

Required methods

pub fn set_speed(&mut self, cr: &mut CR, speed: IOPinSpeed)

Loading content...