Trait stm32f1xx_hal::gpio::PinMode

pub trait PinMode<CR> {
    pub unsafe fn set_mode(cr: &mut CR) -> Self;
}

NOTE: This trait should ideally be private but must be pub in order to avoid complaints from the compiler.

Required methods

pub unsafe fn set_mode(cr: &mut CR) -> Self

Loading content...